Move all option related code to option.go
