+++ client.go
@@ -4,7 +4,6 @@ package bmclib
 
 import (
 	"context"
-	"crypto/x509"
 	"io"
 	"net/http"
 	"sync"
@@ -55,47 +54,6 @@ type providerConfig struct {
 	asrock   asrockrack.Config
 	gofish   redfish.Config
 	intelamt intelamt.Config
-}
-
-// Option for setting optional Client values
-type Option func(*Client)
-
-// WithLogger sets the logger
-func WithLogger(logger logr.Logger) Option {
-	return func(args *Client) { args.Logger = logger }
-}
-
-// WithRegistry sets the Registry
-func WithRegistry(registry *registrar.Registry) Option {
-	return func(args *Client) { args.Registry = registry }
-}
-
-// WithSecureTLS enforces trusted TLS connections, with an optional CA certificate pool.
-// Using this option with an nil pool uses the system CAs.
-func WithSecureTLS(rootCAs *x509.CertPool) Option {
-	return func(args *Client) {
-		args.httpClientSetupFuncs = append(args.httpClientSetupFuncs, httpclient.SecureTLSOption(rootCAs))
-	}
-}
-
-// WithHTTPClient sets an http client
-func WithHTTPClient(c *http.Client) Option {
-	return func(args *Client) {
-		args.httpClient = c
-	}
-}
-
-// WithPerProviderTimeout sets the timeout when interacting with a BMC.
-// This timeout value is applied per provider.
-// When not defined and a context with a timeout is passed to a method, the default timeout
-// will be the context timeout duration divided by the number of providers in the registry,
-// meaning, the len(Client.Registry.Drivers).
-// If this per provider timeout is not defined and no context timeout is defined,
-// the defaultConnectTimeout is used.
-func WithPerProviderTimeout(timeout time.Duration) Option {
-	return func(args *Client) {
-		args.perProviderTimeout = func(context.Context) time.Duration { return timeout }
-	}
 }
 
 // NewClient returns a new Client struct
