+++ option.go
@@ -1,6 +1,56 @@
 package bmclib
 
-import "net/http"
+import (
+	"context"
+	"crypto/x509"
+	"net/http"
+	"time"
+
+	"github.com/bmc-toolbox/bmclib/v2/internal/httpclient"
+	"github.com/go-logr/logr"
+	"github.com/jacobweinstock/registrar"
+)
+
+// Option for setting optional Client values
+type Option func(*Client)
+
+// WithLogger sets the logger
+func WithLogger(logger logr.Logger) Option {
+	return func(args *Client) { args.Logger = logger }
+}
+
+// WithRegistry sets the Registry
+func WithRegistry(registry *registrar.Registry) Option {
+	return func(args *Client) { args.Registry = registry }
+}
+
+// WithSecureTLS enforces trusted TLS connections, with an optional CA certificate pool.
+// Using this option with an nil pool uses the system CAs.
+func WithSecureTLS(rootCAs *x509.CertPool) Option {
+	return func(args *Client) {
+		args.httpClientSetupFuncs = append(args.httpClientSetupFuncs, httpclient.SecureTLSOption(rootCAs))
+	}
+}
+
+// WithHTTPClient sets an http client
+func WithHTTPClient(c *http.Client) Option {
+	return func(args *Client) {
+		args.httpClient = c
+	}
+}
+
+// WithPerProviderTimeout sets the timeout when interacting with a BMC.
+// This timeout value is applied per provider.
+// When not defined and a context with a timeout is passed to a method, the default timeout
+// will be the context timeout duration divided by the number of providers in the registry,
+// meaning, the len(Client.Registry.Drivers).
+// If this per provider timeout is not defined and no context timeout is defined,
+// the defaultConnectTimeout is used.
+func WithPerProviderTimeout(timeout time.Duration) Option {
+	return func(args *Client) {
+		args.perProviderTimeout = func(context.Context) time.Duration { return timeout }
+	}
+}
 
 func WithIpmitoolCipherSuite(cipherSuite int) Option {
 	return func(args *Client) {
